+++ src/App.js
@@ -1,37 +1,16 @@
 import React from 'react';
-
-const todoList = [
-  {
-    id: 1,
-    title: 'Pay Rent',
-  },
-  {
-    id: 2,
-    title: 'Complete Assignment',
-  },
-  {
-    id: 3,
-    title: 'Walk Dog',
-  }
-];
+import AddTodoForm from './AddTodoForm';
+import TodoList from './TodoList';
 
 function App() {
   return (
     <div style={{ textAlign: 'center' }}>
       <h1>Todo List</h1>
 
-      <hr />
-      {/* render the list */}
+      <AddTodoForm />
+      <TodoList />
 
-      <ul>
-        {todoList.map(function(item) {
-          return (
-            <li key={item.id}>
-              <span>{item.title}</span>
-            </li>
-          );
-        })}
-      </ul>
+      
     </div>
   );
 }
